Move storage tests
