--- core/store/src/flat/tests/mod.rs
@@ -1,1 +1,0 @@
-mod storage;
